+++ modules/editable-layers/src/edit-modes/scale-mode.ts
@@ -12,7 +12,7 @@ import {coordEach} from '@turf/meta';
 import turfDistance from '@turf/distance';
 import turfTransformScale from '@turf/transform-scale';
 import {getCoord, getGeom} from '@turf/invariant';
-import {FeatureCollection, Position} from '../utils/geojson-types';
+import {FeatureCollection, Position, SimpleFeatureCollection} from '../utils/geojson-types';
 import {
   ModeProps,
   PointerMoveEvent,
@@ -27,7 +27,7 @@ import {GeoJsonEditMode} from './geojson-edit-mode';
 import {ImmutableFeatureCollection} from './immutable-feature-collection';
 
 export class ScaleMode extends GeoJsonEditMode {
-  _geometryBeingScaled: FeatureCollection | null | undefined;
+  _geometryBeingScaled: SimpleFeatureCollection | null | undefined;
   _selectedEditHandle: EditHandleFeature | null | undefined;
   _cornerGuidePoints: Array<EditHandleFeature> = [];
   _cursor: string | null | undefined;
@@ -36,7 +36,6 @@ export class ScaleMode extends GeoJsonEditMode {
   _isSinglePointGeometrySelected = (geometry: FeatureCollection | null | undefined): boolean => {
     const {features} = geometry || {};
     if (Array.isArray(features) && features.length === 1) {
-      // @ts-expect-error turf types diff
       const {type}: {type: string} = getGeom(features[0]);
       return type === 'Point';
     }
@@ -65,7 +64,7 @@ export class ScaleMode extends GeoJsonEditMode {
     );
   };
 
-  _getUpdatedData = (props: ModeProps<FeatureCollection>, editedData: FeatureCollection) => {
+  _getUpdatedData = (props: ModeProps<SimpleFeatureCollection>, editedData: SimpleFeatureCollection) => {
     let updatedData = new ImmutableFeatureCollection(props.data);
     const selectedIndexes = props.selectedIndexes;
     for (let i = 0; i < selectedIndexes.length; i++) {
@@ -82,7 +81,7 @@ export class ScaleMode extends GeoJsonEditMode {
     startDragPoint: Position,
     currentPoint: Position,
     editType: string,
-    props: ModeProps<FeatureCollection>
+    props: ModeProps<SimpleFeatureCollection>
   ) => {
     if (!this._selectedEditHandle) {
       return null;
@@ -93,9 +92,7 @@ export class ScaleMode extends GeoJsonEditMode {
 
     const scaleFactor = getScaleFactor(origin, startDragPoint, currentPoint);
 
-    // @ts-expect-error turf types diff
-    const scaledFeatures: FeatureCollection = turfTransformScale(
-      // @ts-expect-error turf types diff
+    const scaledFeatures = turfTransformScale(
       this._geometryBeingScaled,
       scaleFactor,
       {origin}
@@ -110,7 +107,7 @@ export class ScaleMode extends GeoJsonEditMode {
     };
   };
 
-  updateCursor = (props: ModeProps<FeatureCollection>) => {
+  updateCursor = (props: ModeProps<SimpleFeatureCollection>) => {
     if (this._selectedEditHandle) {
       if (this._cursor) {
         props.onUpdateCursor(this._cursor);
@@ -118,7 +115,6 @@ export class ScaleMode extends GeoJsonEditMode {
       const cursorGeometry = this.getSelectedFeaturesAsFeatureCollection(props);
 
       // Get resize cursor direction from the hovered scale editHandle (e.g. nesw or nwse)
-      // @ts-expect-error turf types diff
       const centroid = turfCentroid(cursorGeometry);
       const bearing = turfBearing(centroid, this._selectedEditHandle);
       const positiveBearing = bearing < 0 ? bearing + 180 : bearing;
@@ -138,7 +134,7 @@ export class ScaleMode extends GeoJsonEditMode {
     }
   };
 
-  handlePointerMove(event: PointerMoveEvent, props: ModeProps<FeatureCollection>) {
+  handlePointerMove(event: PointerMoveEvent, props: ModeProps<SimpleFeatureCollection>) {
     if (!this._isScaling) {
       const selectedEditHandle = getPickedEditHandle(event.picks);
       this._selectedEditHandle =
@@ -150,7 +146,7 @@ export class ScaleMode extends GeoJsonEditMode {
     }
   }
 
-  handleStartDragging(event: StartDraggingEvent, props: ModeProps<FeatureCollection>) {
+  handleStartDragging(event: StartDraggingEvent, props: ModeProps<SimpleFeatureCollection>) {
     if (this._selectedEditHandle) {
       event.cancelPan();
       this._isScaling = true;
@@ -158,7 +154,7 @@ export class ScaleMode extends GeoJsonEditMode {
     }
   }
 
-  handleDragging(event: DraggingEvent, props: ModeProps<FeatureCollection>) {
+  handleDragging(event: DraggingEvent, props: ModeProps<SimpleFeatureCollection>) {
     if (!this._isScaling) {
       return;
     }
@@ -178,7 +174,7 @@ export class ScaleMode extends GeoJsonEditMode {
     event.cancelPan();
   }
 
-  handleStopDragging(event: StopDraggingEvent, props: ModeProps<FeatureCollection>) {
+  handleStopDragging(event: StopDraggingEvent, props: ModeProps<SimpleFeatureCollection>) {
     if (this._isScaling) {
       // Scale the geometry
       const scaleAction = this.getScaleAction(
@@ -200,7 +196,7 @@ export class ScaleMode extends GeoJsonEditMode {
     }
   }
 
-  getGuides(props: ModeProps<FeatureCollection>): GuideFeatureCollection {
+  getGuides(props: ModeProps<SimpleFeatureCollection>): GuideFeatureCollection {
     this._cornerGuidePoints = [];
     const selectedGeometry = this.getSelectedFeaturesAsFeatureCollection(props);