+++ modules/editable-layers/src/edit-modes/utils.ts
@@ -8,23 +8,21 @@ import destination from '@turf/destination';
 import bearing from '@turf/bearing';
 import pointToLineDistance from '@turf/point-to-line-distance';
 import {flattenEach} from '@turf/meta';
-// import type {MultiLineString} from '@turf/helpers';
 import {point} from '@turf/helpers';
 import {getCoords} from '@turf/invariant';
 import {WebMercatorViewport} from 'viewport-mercator-project';
 import {Viewport, Pick, EditHandleFeature, EditHandleType, StartDraggingEvent} from './types';
 import {
-  Geometry,
+  SimpleGeometry,
   Position,
   Point,
   LineString,
   Polygon,
-  FeatureOf,
-  FeatureWithProps,
-  AnyCoordinates
+  Feature,
+  SimpleGeometryCoordinates
 } from '../utils/geojson-types';
 
-export type NearestPointType = FeatureWithProps<Point, {dist: number; index: number}>;
+export type NearestPointType = Feature<Point, {dist: number; index: number}>;
 
 export function toDeckColor(
   color?: [number, number, number, number] | number,
@@ -119,8 +117,8 @@ export function mix(a: number, b: number, ratio: number): number {
 }
 
 export function nearestPointOnProjectedLine(
-  line: FeatureOf<LineString>,
-  inPoint: FeatureOf<Point>,
+  line: Feature<LineString>,
+  inPoint: Feature<Point>,
   viewport: Viewport
 ): NearestPointType {
   const wmViewport = new WebMercatorViewport(viewport);
@@ -187,8 +185,8 @@ export function nearestPointOnProjectedLine(
 }
 
 export function nearestPointOnLine( // <G extends LineString | MultiLineString>(
-  lines: FeatureOf<LineString>,
-  inPoint: FeatureOf<Point>,
+  lines: Feature<LineString>,
+  inPoint: Feature<Point>,
   viewport?: Viewport
 ): NearestPointType {
   let mercator;
@@ -204,11 +202,9 @@ export function nearestPointOnLine( // <G extends LineString | MultiLineString>(
     return closestPoint;
   }
 
-  // @ts-expect-error TODO
   // eslint-disable-next-line max-statements, complexity
   flattenEach(lines, (line: any) => {
     const coords: any = getCoords(line);
-    // @ts-expect-error TODO
     const pointCoords: any = getCoords(inPoint);
 
     let minDist;
@@ -369,7 +365,7 @@ export function getPickedEditHandles(picks: Pick[] | null | undefined): EditHand
 }
 
 export function getEditHandlesForGeometry(
-  geometry: Geometry,
+  geometry: SimpleGeometry,
   featureIndex: number,
   editHandleType: EditHandleType = 'existing'
 ): EditHandleFeature[] {
@@ -474,7 +470,7 @@ function getEditHandlesForCoordinates(
  * @returns Updated coordinates.
  */
 export function updateRectanglePosition(
-  feature: FeatureOf<Polygon>,
+  feature: Feature<Polygon>,
   editHandleIndex: number,
   coords: Position
 ): Position[][] | null {
@@ -501,9 +497,9 @@ export function updateRectanglePosition(
  * @retuns Transformed coordinates.
  */
 export function mapCoords(
-  coords: AnyCoordinates,
+  coords: SimpleGeometryCoordinates,
   callback: (coords: Position) => Position
-): AnyCoordinates {
+): SimpleGeometryCoordinates {
   if (typeof coords[0] === 'number') {
     if (!isNaN(coords[0]) && isFinite(coords[0])) {
       return callback(coords as Position);