+++ modules/editable-layers/src/mode-handlers/draw-rectangle-handler.ts
@@ -24,7 +24,6 @@ export class DrawRectangleHandler extends TwoClickPolygonHandler {
     const corner1 = clickSequence[0];
     const corner2 = event.mapCoords;
 
-    // @ts-expect-error turf type diff
     this._setTentativeFeature(bboxPolygon([corner1[0], corner1[1], corner2[0], corner2[1]]));
 
     return result;